打印日志go.uber.org/zap v1.27.0

+++ services/dict/dict_service.go
@@ -16,7 +16,7 @@ type DictService interface {
 	CreateDict(dto dto.DictDto) error
 
 	// QueryDictList 查询字典列表
-	QueryDictList(current int, pageSize int) ([]models.Dict, int)
+	QueryDictList(current int, pageSize int) ([]models.Dict, int64)
 
 	// UpdateDict 更新字典
 	UpdateDict(userDto dto.DictDto) error
