+++ services/log/log_service.go
@@ -16,7 +16,7 @@ type LogService interface {
 	CreateLog(dto dto.LogDto) error
 
 	// QueryLogList 查询操作日志
-	QueryLogList(current int, pageSize int) ([]models.OperationLog, int)
+	QueryLogList(current int, pageSize int) ([]models.OperationLog, int64)
 
 	// DeleteLogByIds 删除操作日志
 	DeleteLogByIds(ids []int64) error
@@ -25,7 +25,7 @@ type LogService interface {
 	CreateLoginLog(dto dto.LoginLogDto) error
 
 	// QueryLoginLogList 查询登录日志
-	QueryLoginLogList(current int, pageSize int) ([]models.LoginLog, int)
+	QueryLoginLogList(current int, pageSize int) ([]models.LoginLog, int64)
 
 	// DeleteLoginLogByIds 删除登录日志
 	DeleteLoginLogByIds(ids []int64) error
