+++ services/log/log_service_impl.go
@@ -25,7 +25,7 @@ func (l *LogServiceImpl) CreateLog(dto dto.LogDto) error {
 }
 
 // QueryLogList 查询操作日志
-func (l *LogServiceImpl) QueryLogList(current int, pageSize int) ([]models.OperationLog, int) {
+func (l *LogServiceImpl) QueryLogList(current int, pageSize int) ([]models.OperationLog, int64) {
 	return l.Dao.LogDao.QueryLogList(current, pageSize)
 }
 
@@ -40,7 +40,7 @@ func (l *LogServiceImpl) CreateLoginLog(dto dto.LoginLogDto) error {
 }
 
 // QueryLoginLogList 查询登录日志
-func (l *LogServiceImpl) QueryLoginLogList(current int, pageSize int) ([]models.LoginLog, int) {
+func (l *LogServiceImpl) QueryLoginLogList(current int, pageSize int) ([]models.LoginLog, int64) {
 	return l.Dao.LogDao.QueryLoginLogList(current, pageSize)
 }
