+++ services/menu/menu_service_impl.go
@@ -25,7 +25,7 @@ func (menu *MenuServiceImpl) CreateMenu(dto dto.MenuDto) error {
 }
 
 // QueryMenuList 查询菜单
-func (menu *MenuServiceImpl) QueryMenuList(username string) ([]models.Menu, int) {
+func (menu *MenuServiceImpl) QueryMenuList(username string) ([]models.Menu, int64) {
 	return menu.Dao.MenuDao.QueryMenuList()
 }
